requirement.txt
streamlit
httpx
python-dotenv
fastapi
uvicorn[standard]
python-multipart